Manage dinosaurs_count in cage model
Avoid doing expensive dinosaur count after each cage save

+++ app/services/dinosaurs/delete_service.rb
@@ -6,9 +6,18 @@ module Dinosaurs
 
     def call
       dinosaur = Dinosaur.find(params[:id])
-      dinosaur.destroy
 
-      Success()
+      ActiveRecord::Base.transaction(requires_new: true) do
+        cage = dinosaur.cage
+        if cage.present?
+          cage.dinosaurs&.delete(dinosaur)
+          return resource_failure(cage) unless cage.save
+        end
+
+        dinosaur.destroy
+
+        Success()
+      end
     rescue StandardError
       resource_failure(dinosaur)
     end
